refactor: join api 수정
