+++ src/users/users.module.ts
@@ -3,12 +3,11 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Users } from './entities/users.entity';
-import { UserRepository } from './users.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Users])],
   controllers: [UsersController],
-  providers: [UsersService, UserRepository],
+  providers: [UsersService],
   exports: [UsersService],
 })
 export class UsersModule {}
